history: Remove port history data in the background
 * persist/unqlite: Remove redundant filter matching call
 * utils: Move ThreadedRunner to utils/asyncio
 * core/history: Remove data in the background
